add a function to get best tuning config
